change size_t to unsigned int.
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@4750 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ highlevel/highlevel_controllers/src/move_base_sbpl.cpp
@@ -199,7 +199,7 @@ namespace ros {
       // Apply map updates that were buffered from the cost map
       applyMapUpdates();
 
-      size_t x, y;
+      unsigned int x, y;
       const CostMap2D& cm = getCostMap();
 
       // Set start state based on global pose.
